requirements.txt
fastapi
pandas
openai
pypdf
numpy
openai-agent
chromadb
langchain
langchain-community
langchain-core
python-dotenv
langchain-groq
wikipedia

